requirements.txt
APScheduler==3.10.1
dicttoxml==1.7.16
Flask==2.2.3
Flask_json_schema==0.0.5
PyYAML==6.0
Requests==2.28.2
